test code for assembly
